Move image classes to new module
