+++ plantcv/plantcv/__init__.py
@@ -5,7 +5,7 @@ from plantcv.plantcv.classes import Outputs
 from plantcv.plantcv.classes import Spectral_data
 from plantcv.plantcv.classes import PSII_data
 from plantcv.plantcv.classes import Points
-from plantcv.plantcv.classes import Image, BGR, RGB, GRAY, HSI
+from plantcv.plantcv.images import Image, BGR, RGB, GRAY, HSI
 
 # Initialize an instance of the Params and Outputs class with default values
 # params and outputs are available when plantcv is imported
